refactor: move health endpoints to dedicated health module
- Create app/health/ module with router.py and checks.py
- Move health check logic from app/health.py to app/health/checks.py
- Move health endpoints from main.py to dedicated health router
- Clean up imports and unused dependencies in main.py
- Maintain same endpoint paths for Kubernetes compatibility
- Improve code organization following FastAPI best practices
