+++ app/health/__init__.py
@@ -1,0 +1,2 @@
+# Health module
+from .checks import mark_startup_complete
